add --resume opt

+++ veriLLMs/verithoughts/verilog_openai.py
@@ -40,12 +40,14 @@ parser.add_argument("--enable_reasoning", action="store_true", help="Enable if y
 parser.add_argument("--temperature", type=float, default=0.6, help="Temperature")
 parser.add_argument("--top_p", type=float, default=0.95, help="Top p")
 parser.add_argument("--max_tokens", type=int, default=32768, help="Max tokens") # Not used
+parser.add_argument("--resume_gen", action="store_true", help="Enable if you want to continue from existing results.jsonl file")
 args = parser.parse_args()
 
-# # based on what's hardcoded in verilog_vllm (their repo)!
-# temperature=0.6
-# top_p = 0.95
-# max_tokens=32768
+temperature=args.temperature
+top_p = args.top_p
+max_tokens=args.max_tokens
+resume_gen=args.resume_gen
+
 
 model_name = args.model_name
 num_samples_per_task = args.num_samples_per_task
@@ -60,8 +62,15 @@ benchmark_data = load_dataset("wilyub/VeriThoughtsBenchmark", split="train")
 results_path = os.path.join("benchmark_results", model_name)
 os.makedirs(results_path, exist_ok=True)
 results_file = os.path.join(results_path, "results.jsonl")
-with open(results_file, "w") as f:
-    pass # reset! their code appends indefinitely! OMG!
+if resume_gen and os.path.exists(results_file):
+    with open(results_file, "r") as rf:
+        existing_results = json.load(rf)
+    already_done = len(existing_results)
+else:
+    already_done = 0
+    with open(results_file, "w") as f:
+        pass # reset! their code appends indefinitely! OMG!
+
 
 INSTR_SIMPLE = "Make sure your input and output interface has the same names as described in the question. \nPlease start your Verilog code with CODE BEGIN and end with CODE END.\n" 
 INSTR_REASONING = "Make sure your input and output interface has the same names as described in the question. \nPlease start your Verilog code with CODE BEGIN and end with CODE END.<think>\n"
@@ -79,6 +88,8 @@ for data in benchmark_data:
 
 
 for i, question in enumerate(tqdm(question_list, desc="Processing VeriThought questions")):
+    if i < already_done:
+        continue
 
     benchmark_dict = verified_benchmark_dict_list[i]
     question = question_list[i]
